BAC-39: create simple cli managment program for dev

+++ app/models/__init__.py
@@ -10,8 +10,9 @@ from .Product import Product
 from .User import User
 
 
-async def recreate_all_tables():
+async def create_all_tables(drop_all=False):
     """Drops all tables and then recreates them"""
     async with engine.begin() as conn:
-        await conn.run_sync(Base.metadata.drop_all)
+        if drop_all:
+            await conn.run_sync(Base.metadata.drop_all)
         await conn.run_sync(Base.metadata.create_all)
